+++ packages/band/src/quantiles/index.tsx
@@ -1,5 +1,6 @@
 export * from './context'
 export * from './Quantile'
+export * from './QuantileTooltip'
 export * from './Quantiles'
 export * from './BoxAndWhiskers'
 export * from './predicates'