+++ packages/core/tests/typography/Counter.test.tsx
@@ -1,4 +1,4 @@
-import { fireEvent, render, screen, waitFor } from '@testing-library/react'
+import { fireEvent, render, screen } from '@testing-library/react'
 import { Chart, Counter } from '../../src'
 import { chartProps } from '../props'
 import { useState } from 'react'