+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+-   [band] custom tooltip component `QuantileTooltip` to display quantile
+    details in a table
 -   [xy] support for error bars with `ScatterErrors`
 -   [examples] random walks with moving averages
 -   [xy] basic signal-processing props (convolution, downsampling) for curves,